MAINT: Add Checks To User Provided Paths (#322)
- Address Deepsource security issues 
- Avoid the user from accidentally overwriting files or writing to a directory that doesn't exists.

+++ tiatoolbox/utils/misc.py
@@ -94,12 +94,25 @@ def grab_files_from_dir(input_path, file_types=("*.jpg", "*.png", "*.tif")):
     return list(files_grabbed)
 
 
-def save_yaml(input_dict, output_path="output.yaml"):
+def save_yaml(
+    input_dict: dict,
+    output_path="output.yaml",
+    parents: bool = False,
+    exist_ok: bool = False,
+):
     """Save dictionary as yaml.
 
     Args:
-        input_dict (dict): A variable of type 'dict'
-        output_path (str or pathlib.Path): Path to save the output file
+        input_dict (dict):
+            A variable of type 'dict'
+        output_path (str or pathlib.Path):
+            Path to save the output file
+        parents (bool):
+            Make parent directories if they do not exist. Default is
+            False.
+        exist_ok (bool):
+            Overwrite the output file if it exists. Default is False.
+
 
     Returns:
 
@@ -109,7 +122,14 @@ def save_yaml(input_dict, output_path="output.yaml"):
         >>> utils.misc.save_yaml(input_dict, './hello.yaml')
 
     """
-    with open(str(pathlib.Path(output_path)), "w") as yaml_file:
+    path = pathlib.Path(output_path)
+    if path.exists() and not exist_ok:
+        raise FileExistsError("File already exists.")
+    if parents:
+        path.parent.mkdir(parents=True, exist_ok=True)
+    with open(  # skipcq: PTC-W6004: PTC-W6004
+        str(pathlib.Path(output_path)), "w"
+    ) as yaml_file:
         yaml.dump(input_dict, yaml_file)
 
 
@@ -660,7 +680,12 @@ def __walk_dict(dct):
         dct[k] = __walk_list_dict(v)
 
 
-def save_as_json(data, save_path):
+def save_as_json(
+    data: Union[dict, list],
+    save_path: Union[str, pathlib.Path],
+    parents: bool = False,
+    exist_ok: bool = False,
+):
     """Save data to a json file.
 
     The function will deepcopy the `data` and then jsonify the content
@@ -668,8 +693,16 @@ def save_as_json(data, save_path):
     `bool` and their np.ndarray respectively.
 
     Args:
-        data (dict or list): Input data to save.
-        save_path (str): Output to save the json of `input`.
+        data (dict or list):
+            Input data to save.
+        save_path (str):
+            Output to save the json of `input`.
+        parents (bool):
+            Make parent directories if they do not exist. Default is
+            False.
+        exist_ok (bool):
+            Overwrite the output file if it exists. Default is False.
+
 
     """
     shadow_data = copy.deepcopy(data)  # make a copy of source input
@@ -681,7 +714,12 @@ def save_as_json(data, save_path):
     else:
         __walk_list(shadow_data)
 
-    with open(save_path, "w") as handle:
+    save_path = pathlib.Path(save_path)
+    if save_path.exists() and not exist_ok:
+        raise FileExistsError("File already exists.")
+    if parents:
+        save_path.parent.mkdir(parents=True, exist_ok=True)
+    with open(save_path, "w") as handle:  # skipcq: PTC-W6004: PTC-W6004
         json.dump(shadow_data, handle)
